Update core streaming classes to support R2020b
R2020b changed the access permissions for a number of core classes which
broken all streaming interface classes. All classes have been updated to
fix this and a test was added to make sure all classes initiate
correctly. Submodule for common pieces was updated as well which
supports R2020b.

Signed-off-by: Travis F. Collins <travis.collins@analog.com>